Change MC years weight to float instead of int

+++ src/solver/variable/storage/stdDeviation.h
@@ -294,8 +294,8 @@ namespace AllYears
 
 	private:
 
-        std::vector<int>    yearsWeight;
-        int                 yearsWeightSum;
+        std::vector<float>    yearsWeight;
+		float                 yearsWeightSum;
         
 	}; // class StdDeviation
